+++ apps/omsorgspengesoknad/package.json
@@ -69,6 +69,7 @@
         "axe-core": "4.10.0",
         "babel-polyfill": "6.26.0",
         "busboy": "1.6.0",
+        "chromatic": "^11.8.0",
         "cors": "2.8.5",
         "cross-env": "7.0.3",
         "helmet": "7.1.0",
