+++ apps/omsorgspengesoknad/src/app/i18n/app-messages.stories.tsx
@@ -1,17 +1,16 @@
 import { Meta, StoryFn } from '@storybook/react';
-import * as React from 'react';
 import MessagesPreview from '@navikt/sif-common-core-ds/src/dev-utils/intl/messages-preview/MessagesPreview';
-import { libMessages } from '../../../app/i18n/index';
-import KvitteringPage from '../../../app/pages/kvittering/KvitteringPage';
-import { withIntl } from '../../decorators/withIntl';
+import KvitteringPage from '../pages/kvittering/KvitteringPage';
+import { withIntl } from '../../storybook/decorators/withIntl';
+import { appMessages } from './appMessages';
 
 export default {
-    title: 'i18N/Lib tekster',
+    title: 'i18N/App-tekster',
     decorators: [withIntl],
 } as Meta<typeof KvitteringPage>;
 
 const Template: StoryFn = () => (
-    <MessagesPreview showExplanation={false} messages={libMessages} showMissingTextSummary={true} />
+    <MessagesPreview showExplanation={false} messages={appMessages} showMissingTextSummary={true} />
 );
 export const Default = Template.bind({});
