+++ apps/omsorgspengesoknad/src/app/pages/velkommen/VelkommenPage.stories.tsx
@@ -1,17 +1,16 @@
 import { Meta, StoryFn } from '@storybook/react';
-import * as React from 'react';
-import VelkommenPage from '../../../app/pages/velkommen/VelkommenPage';
-import { withIntl } from '../../decorators/withIntl';
-import { withRouterProvider } from '../../decorators/withRouter';
-import { withAmplitudeProvider } from '../../decorators/withAmplitudeProvider';
-import { withSøknadContextProvider } from '../../decorators/withSøknadContext';
-import { velkommenPageMessages } from '../../../app/pages/velkommen/velkommenPageMessages';
+import VelkommenPage from './VelkommenPage';
+import { withIntl } from '../../../storybook/decorators/withIntl';
+import { withRouterProvider } from '../../../storybook/decorators/withRouter';
+import { withAmplitudeProvider } from '../../../storybook/decorators/withAmplitudeProvider';
+import { withSøknadContextProvider } from '../../../storybook/decorators/withSøknadContext';
+import { velkommenPageMessages } from './velkommenPageMessages';
 import MessagesPreview from '@navikt/sif-common-core-ds/src/dev-utils/intl/messages-preview/MessagesPreview';
 import { Tabs } from '@navikt/ds-react';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 
 export default {
-    title: 'Pages/VelkommenPage',
+    title: 'Pages/Velkommen',
     component: VelkommenPage,
     decorators: [withIntl, withRouterProvider, withAmplitudeProvider, withSøknadContextProvider],
 } as Meta<typeof VelkommenPage>;
