+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/alert/OmBarnetAlerts.stories.tsx
@@ -8,7 +8,7 @@ import TrengerIkkeSøkeForBarnAlert from './TrengerIkkeSøkeForBarnAlert';
 import AlertStoryWrapper from '../../../../../storybook/components/AlertStoryWrapper';
 
 export default {
-    title: 'Steg/OmBarnet/Alerts',
+    title: 'Steps/OmBarnet/Alerts',
     decorators: [withIntl],
 };
