+++ apps/omsorgspengesoknad/src/storybook/utils/intlUtils.ts
@@ -1,7 +1,7 @@
 import { appMessages } from '../../app/i18n/appMessages';
 import { MessageFileFormat } from '@navikt/sif-common-core-ds/src/types';
 
-export const getIntlMessagesFromKeys = (keys: string[]): MessageFileFormat => {
+const getIntlMessagesFromKeys = (keys: string[]): MessageFileFormat => {
     const messages: MessageFileFormat = {
         nb: {},
         nn: {},
@@ -14,14 +14,20 @@ export const getIntlMessagesFromKeys = (keys: string[]): MessageFileFormat => {
     return messages;
 };
 
-export const getValidationIntlKeys = (keys: string[], intlKeyPath: string) => {
+const getValidationIntlKeys = (keys: string[], intlKeyPath: string) => {
     return keys.map((key) => `${intlKeyPath}.${key}`);
 };
 
-export const getScopedIntlKeys = (scope: string): string[] => {
+const getScopedIntlKeys = (scope: string): string[] => {
     return Object.keys(appMessages.nb).filter((key) => key.includes(scope));
 };
 
-export const includesString = (str: string, searchFor: string | string[]): boolean => {
-    return Array.isArray(searchFor) ? searchFor.some((subStr) => str.includes(subStr)) : str.includes(searchFor);
+// const includesString = (str: string, searchFor: string | string[]): boolean => {
+//     return Array.isArray(searchFor) ? searchFor.some((subStr) => str.includes(subStr)) : str.includes(searchFor);
+// };
+
+export const storybookIntlUtils = {
+    getIntlMessagesFromKeys,
+    getScopedIntlKeys,
+    getValidationIntlKeys,
 };
