+++ apps/omsorgspengesoknad/tsconfig.json
@@ -5,5 +5,5 @@
         "typeRoots": ["../../node_modules/@types", "./node_modules/@types"]
     },
     "exclude": ["node_modules", "src/build", "**.js", "dist"],
-    "include": ["./src/app/**/*"]
+    "include": ["./src/app/**/*", "./src/storybook/**/*"]
 }
